moved server>chat.js to main server js file
