+++ package.json
@@ -40,6 +40,7 @@
   "dependencies": {
     "alert": "^5.0.10",
     "bcrypt": "^5.0.0",
+    "bootstrap": "^5.0.2",
     "connect-session-sequelize": "^7.0.3",
     "cookie-parser": "^1.4.5",
     "dotenv": "^8.6.0",
@@ -57,9 +58,9 @@
     "passport-facebook": "^3.0.0",
     "passport-google": "^0.3.0",
     "passport-local": "^1.0.0",
-    "socket.io": "^4.1.2",
     "passport-twitter": "^1.0.4",
     "sequelize": "^6.6.4",
+    "socket.io": "^4.1.2",
     "util": "^0.12.4"
   },
   "devDependencies": {
