+++ public/assets/css/style.css
@@ -1,8 +1,3 @@
-body {
-  color: white;
-  background-size: cover;
-  height:100%
-}
 
 .jumbotron {
   background: rgba(0,0,0,0);
