+++ public/chat.html
@@ -1,26 +1,90 @@
 <!doctype html>
 <html>
-    <head>
-        <title>💬 Real-Time Chat App with Socket.IO</title>
-        <meta charset="UTF-8" />
-        <link rel="stylesheet" href="style.css" />
-    </head>
-    <body>
-        <div class="main">
-        <h1>Join the conversation</h1>
-        <ul class="feed"></ul>
+
+<head>
+    <title>💬 Real-Time Chat App with Socket.IO</title>
+    <meta charset="UTF-8" />
+    <link rel="stylesheet" href="style.css" />
+
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
+</head>
+
+<body>
+  <nav class="navbar navbar-expand-lg navbar-dark ">
+    <a class="navbar-brand" href="#">GeoGuessr</a>
+    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+      <span class="navbar-toggler-icon"></span>
+    </button>
+  
+    <div class="collapse navbar-collapse" id="navbarSupportedContent">
+      <ul class="navbar-nav mr-auto">
         
-        <form class="messaging-form" action="#">
-            <span class="feedback"></span>
-            <div class="message-input">
-                <span class="avatar">?</span>
-                <input type="text" class="message-input-field name-input" placeholder="Your name" />
+        <li class="nav-item dropdown">
+          <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+            Dropdown
+          </a>
+          <div class="dropdown-menu" aria-labelledby="navbarDropdown">
+            <a class="dropdown-item" href="#">Action</a>
+            <a class="dropdown-item" href="#">Another action</a>
+            <div class="dropdown-divider"></div>
+            <a class="dropdown-item" href="#">Something else here</a>
+          </div>
+        </li>
+        <div class="timer">Time Remaining : 60</div>
+    </div>
+  </nav>
+
+      <section class="geobox">
+
+      </section>
+
+    <div class="container-fluid">
+        <div class="row">
+            <div class="col-md-8">
+            <img class="picclass" src="/public/pics/Google-Street-View-overlay-2.jpg" alt="">
             </div>
-            
-            <button class="join">Join</button>
-        </form>
+            <div class="col-md-3 offset-md-1 ">
+                <div class="main row align-items-end">
+                    <h1>Join the conversation</h1>
+                    <ul class="feed"></ul>
+
+                    <form class="messaging-form" action="#">
+                        <span class="feedback"></span>
+                        <div class="message-input">
+                            <span class="avatar">?</span>
+                            <button class="join buttun">Join</button>
+                            <input type="text" class="message-input-field name-input" placeholder="Your name" />
+                        </div>
+
+                    </form>
+
+                </div>
+            </div>
+        </div>
+    </div>
+    <div class="container-fluid footer">
+      <div class="row">
+          <div class="col-md-2">
+            1st place : Tom
+          </div>
+          <div class="col-md-2">
+            2nd place : Burke
+         </div>
+         <div class="col-md-2">
+         3rd place : Footlover
+       </div>
+          </div>
+          </div>
+
+  
 </div>
 </body>
+<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
+<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
 <script src="https://cdn.socket.io/socket.io-1.2.0.js"></script>
 <script src="app.js"></script>
+
 </html>
+
