+++ public/style.css
@@ -1,14 +1,29 @@
-
 ::placeholder {
     font-style: italic;
 }
-
-
-
-body {
-    margin: 30px;
+.navbar-dark {
+   
     background: #252535;
-    font-family: sans-serif;
+    
+}
+
+.footer {
+    position: fixed;
+    left: 0;
+    bottom: 0;
+    width: 100%;
+    background-color: #252535;
+    color: white;
+    text-align: center;
+    z-index: -100;
+    height: 6%;
+  }
+.navbar-nav {
+    display: flex;
+    flex-direction: column;
+    padding-left: 0;
+    margin-bottom: 0;
+    list-style: none;
 }
 
 ul {
@@ -32,12 +47,23 @@ h1 {
     margin: 0;
 }
 
+.geobox{
+    width: 66%;
+    height: 100%;
+    background-color: #212121;
+}
+.chatbox{
+background-color:#212121;
+}
+
 time {
-    font-size: 70%;
+    font-size: 90%;
+    color: #212121;
 }
 
 p {
     margin: 10px 0 0;
+    color: #212121;
 }
 
 hr {
@@ -54,14 +80,9 @@ hr:after {
 }
 
 .messaging-form {
+   
     position: fixed;
-    left: 0;
     bottom: 0;
-    padding: 25px 20px;
-    background: #252535;
-    width: 100%;
-    box-sizing: border-box;
-    display: flex;
 }
 
 .message-input {
@@ -70,9 +91,15 @@ hr:after {
     padding: 2px;
     border-radius: 40px;
     margin-right: 20px;
-    background:white;
+    
 }
-
+.timer{
+    color: white;
+    position: fixed;
+    right: 0;
+    margin-right: 20px;
+    margin-top: 10px;
+}
 .avatar {
     width: 25px;
     height: 25px;
@@ -100,10 +127,13 @@ hr:after {
     background: #252535;
     color: #FFF;
     outline: none;
+    height: 10%;
 }
 
-.message-input.error {
-    background: red;
+.picclass {
+   width: 100%;
+   height: 100%;
+    
 }
 
 .messaging-form .message-input-field:focus {
@@ -111,7 +141,7 @@ hr:after {
 }
 
 .join {
-    background: #ebeeec;
+    background: white;
     border: 0;
     color: #252535;
     font-weight: bold;
@@ -123,7 +153,7 @@ hr:after {
 
 .welcome-message {
     text-align: center;
-    color: #fafffc;
+    color: #212121;
     margin: 40px 0;
 }
 
@@ -189,5 +219,5 @@ hr:after {
 }
 
 .message-entry-own .user-name {
-    color: #858282;
+    color: #212121;
 }
