+++ server.js
@@ -11,7 +11,11 @@ const app = express()
 const db = require('./models')
 const uploadRoutes = require('./routes/uploadRoutes')
 const initRoutes = require('./routes/web')
-
+const http = require('http');
+const { Server } = require("socket.io");
+const server = http.createServer(app);
+const io = new Server(server);
+const path = require('path')
 global.__basedir = __dirname
 initRoutes(app)
 
@@ -69,3 +73,48 @@ db.sequelize.sync(syncOptions).then(() => {
 app.use('/upload', uploadRoutes)
 
 module.exports = app
+
+
+
+
+const users = {}
+const typers = {}
+
+io.on('connection', socket => {
+  console.log('connected...')
+
+  socket.on('user connected', payload => {
+    users[socket.id] = {
+      id: socket.id,
+      name: payload.name,
+      avatar: payload.avatar
+    }
+
+    socket.broadcast.emit('user connected', users[socket.id])
+  })
+
+  socket.on('user typing', () => {
+    typers[socket.id] = 1
+
+    socket.broadcast.emit('user typing', {
+      user: users[socket.id].name,
+      typers: Object.keys(typers).length
+    })
+  })
+
+  socket.on('user stopped typing', () => {
+    delete typers[socket.id]
+
+    socket.broadcast.emit('user stopped typing', Object.keys(typers).length)
+  })
+
+  socket.on('send message', payload => {
+    delete typers[socket.id]
+
+    socket.broadcast.emit('send message', {
+      user: payload.user,
+      message: payload.message,
+      typers: Object.keys(typers).length
+    })
+  })
+})
